+++ src/main/kotlin/com/musinsa/musinsacodingassignment/category/domain/Category.kt
@@ -1,7 +1,5 @@
 package com.musinsa.musinsacodingassignment.category.domain
 
-import com.musinsa.musinsacodingassignment.brand.entity.CategoryEntity
-
 data class Category(
     var id: Long?,
     var name: String
@@ -11,13 +9,3 @@ data class Category(
         return this
     }
 }
-
-fun Category.toEntity() = CategoryEntity(
-    id = this.id ?: 0,
-    name = this.name
-)
-
-fun CategoryEntity.toDomain() = Category(
-    id = this.id,
-    name = this.name,
-)
