+++ src/main/kotlin/com/musinsa/musinsacodingassignment/category/entity/CategoryEntity.kt
@@ -1,5 +1,6 @@
-package com.musinsa.musinsacodingassignment.brand.entity
+package com.musinsa.musinsacodingassignment.category.entity
 
+import com.musinsa.musinsacodingassignment.category.domain.Category
 import com.musinsa.musinsacodingassignment.common.domain.BaseEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
@@ -15,3 +16,13 @@ class CategoryEntity(
         this.id = id
     }
 }
+
+fun Category.toEntity() = CategoryEntity(
+    id = this.id ?: 0,
+    name = this.name
+)
+
+fun CategoryEntity.toDomain() = Category(
+    id = this.id,
+    name = this.name,
+)
