+++ src/main/kotlin/com/musinsa/musinsacodingassignment/category/repository/CategoryRepository.kt
@@ -1,8 +1,10 @@
 package com.musinsa.musinsacodingassignment.category.repository
 
-import com.musinsa.musinsacodingassignment.brand.entity.CategoryEntity
+import com.musinsa.musinsacodingassignment.category.entity.CategoryEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CategoryRepository : JpaRepository<CategoryEntity, Long> {
     fun findByName(name: String): CategoryEntity?
+
+    fun existsByName(name: String): Boolean
 }
